refactor : index export에 따른 경로 수정
