+++ src/app/(sub)/join/funnels/step1/components/NumberForm.client.tsx
@@ -5,7 +5,7 @@ import { formatNumber, formatNumberBackSpace } from '../util';
 import { useSMSMutation } from '@/apis/auth';
 import { Button, ButtonGroup } from '@/components/Button';
 import { Spacing } from '@/components/common/Spacing';
-import { useTimerContext } from '@/components/Provider/TimerProvider.client';
+import { useTimerContext } from '@/components/Provider';
 import { TextFieldController } from '@/components/TextField';
 import { regexr } from '@/constants/regexr';
 import { ElementType, KeyboardEventHandler } from 'react';
